requirements.txt
# Python dependencies for Gem5 Simulation Pipeline

# Core dependencies
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# Google Drive integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Progress tracking and monitoring
tqdm>=4.66.0
rich>=13.5.0

# Logging and utilities
python-dateutil>=2.8.2
psutil>=5.9.0

# Configuration management
jsonschema>=4.19.0

# Optional: for advanced data processing
scikit-learn>=1.3.0  # For future ML integration
matplotlib>=3.7.0    # For visualization
seaborn>=0.12.0      # For advanced plotting
